+++ src/main/java/com/comp2024b/tocountornot/service/CategoryService.java
@@ -1,7 +1,9 @@
 package com.comp2024b.tocountornot.service;
 
+import com.comp2024b.tocountornot.bean.Bill;
 import com.comp2024b.tocountornot.bean.Category;
 import com.comp2024b.tocountornot.dao.CategoryMapper;
+import com.comp2024b.tocountornot.exception.ForbiddenException;
 import com.comp2024b.tocountornot.exception.NotFoundException;
 import org.springframework.stereotype.Service;
 
@@ -21,7 +23,11 @@ public class CategoryService {
 
     public void deleteCategory(int id, int uid) {
         if (ExistCategory(id,uid)) {
-            categoryMapper.deleteCategory(id);
+            if (!ExistBill(id)) {
+                categoryMapper.deleteCategory(id);
+            } else {
+                throw new ForbiddenException("cannot delete category cause there is at least a bill under it");
+            }
         } else {
             throw new NotFoundException("category not found");
         }
@@ -29,7 +35,12 @@ public class CategoryService {
 
     public void updateCategory(Category category, int uid) {
         if (ExistCategory(category.getId(),uid)) {
-            categoryMapper.updateCategory(category);
+            Category c = getCategoryById(category.getId(), uid);
+            if (c.getDivision() == category.getDivision()) {
+                categoryMapper.updateCategory(category);
+            } else {
+                throw new ForbiddenException("cannot modify division");
+            }
         } else {
             throw new NotFoundException("category not found");
         }
@@ -51,5 +62,10 @@ public class CategoryService {
     public boolean ExistCategory(int id, int uid) {
         Category category = getCategoryById(id, uid);
         return category != null;
+    }
+
+    public boolean ExistBill(int id) {
+        List<Bill> list = categoryMapper.getBillByCategory(id);
+        return list != null;
     }
 }
