+++ src/main/java/com/comp2024b/tocountornot/service/MemberService.java
@@ -1,7 +1,9 @@
 package com.comp2024b.tocountornot.service;
 
+import com.comp2024b.tocountornot.bean.Bill;
 import com.comp2024b.tocountornot.bean.Member;
 import com.comp2024b.tocountornot.dao.MemberMapper;
+import com.comp2024b.tocountornot.exception.ForbiddenException;
 import com.comp2024b.tocountornot.exception.NotFoundException;
 import org.springframework.stereotype.Service;
 
@@ -21,7 +23,11 @@ public class MemberService {
 
     public void deleteMember(int id, int uid) {
         if (ExistMember(id, uid)) {
-            memberMapper.deleteMember(id);
+            if (!ExistBill(id)) {
+                memberMapper.deleteMember(id);
+            } else {
+                throw new ForbiddenException("cannot delete member cause there is at least a bill under it");
+            }
         } else {
             throw new NotFoundException("member not found");
         }
@@ -51,5 +57,10 @@ public class MemberService {
     public boolean ExistMember(int id, int uid) {
         Member member = getMemberById(id, uid);
         return member != null;
+    }
+
+    public boolean ExistBill(int id) {
+        List<Bill> list = memberMapper.getBillByMember(id);
+        return list != null;
     }
 }
